# restful/docs/swagger.yaml
definitions:
...
  models.RoutineAccessToken:
    properties:
      accessToken:
        description: openid
        type: string
      id:
        description: 小程序access_token表ID
        type: integer
      stopTime:
        description: 添加时间
        type: integer
    type: object
  models.RoutineFormId:
    properties:
      formId:
        description: 表单ID
        type: string
      id:
        description: 表单ID表ID
        type: integer
      status:
        description: 状态1 未使用 2不能使用
        type: integer
      stopTime:
        description: 表单ID失效时间
        type: integer
      uid:
        description: 用户uid
        type: integer
    type: object
...
  models.User:
    properties:
      account:
        description: 用户账号
        type: string
      addIp:
        description: 添加ip
        type: string
      addTime:
        description: 添加时间
        type: integer
      addres:
        description: 详细地址
        type: string
      avatar:
        description: 用户头像
        type: string
      birthday:
        description: 生日
        type: integer
      cardId:
        description: 身份证号码
        type: string
      cleanTime:
        description: 等级清除时间为0没有清除过
        type: integer
      groupId:
        description: 用户分组id
        type: integer
      integral:
        description: 用户剩余积分
        type: number
      isPromoter:
        description: 是否为推广员
        type: integer
      lastIp:
        description: 最后一次登录ip
        type: string
      lastTime:
        description: 最后一次登录时间
        type: integer
      level:
        description: 等级
        type: integer
      mark:
        description: 用户备注
        type: string
      nickname:
        description: 用户昵称
        type: string
      nowMoney:
        description: 用户余额
        type: number
      partnerId:
        description: 合伙人id
        type: integer
      payCount:
        description: 用户购买次数
        type: integer
      phone:
        description: 手机号码
        type: string
      pwd:
        description: 用户密码
        type: string
      realName:
        description: 真实姓名
        type: string
      signNum:
        description: 连续签到天数
        type: integer
      spreadCount:
        description: 下级人数
        type: integer
      spreadTime:
        description: 推广员关联时间
        type: integer
      spreadUid:
        description: 推广元id
        type: integer
      status:
        description: 1为正常，0为禁止
        type: integer
      uid:
        description: 用户id
        type: integer
      userType:
        description: 用户类型
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /routineAccessTokens:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: routineAccessToken array
          schema:
            items:
              $ref: '#/definitions/models.RoutineAccessToken'
            type: array
      summary: List
      tags:
      - RoutineAccessToken
    post:
      parameters:
      - description: RoutineAccessToken
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RoutineAccessToken'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - RoutineAccessToken
  /routineAccessTokens/{routineAccessTokenId}:
    delete:
      parameters:
      - description: routineAccessTokenId
        in: path
        name: routineAccessTokenId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - RoutineAccessToken
    get:
      parameters:
      - description: routineAccessTokenId
        in: path
        name: routineAccessTokenId
        required: true
        type: string
      responses:
        "200":
          description: routineAccessToken object
          schema:
            $ref: '#/definitions/models.RoutineAccessToken'
      summary: Get
      tags:
      - RoutineAccessToken
    patch:
      parameters:
      - description: routineAccessToken
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RoutineAccessToken'
      - description: routineAccessTokenId
        in: path
        name: routineAccessTokenId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - RoutineAccessToken
    put:
      parameters:
      - description: routineAccessToken
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RoutineAccessToken'
      - description: routineAccessTokenId
        in: path
        name: routineAccessTokenId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - RoutineAccessToken
  /routineFormIds:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: routineFormId array
          schema:
            items:
              $ref: '#/definitions/models.RoutineFormId'
            type: array
      summary: List
      tags:
      - RoutineFormId
    post:
      parameters:
      - description: RoutineFormId
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RoutineFormId'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - RoutineFormId
  /routineFormIds/{routineFormIdId}:
    delete:
      parameters:
      - description: routineFormIdId
        in: path
        name: routineFormIdId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - RoutineFormId
    get:
      parameters:
      - description: routineFormIdId
        in: path
        name: routineFormIdId
        required: true
        type: string
      responses:
        "200":
          description: routineFormId object
          schema:
            $ref: '#/definitions/models.RoutineFormId'
      summary: Get
      tags:
      - RoutineFormId
    patch:
      parameters:
      - description: routineFormId
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RoutineFormId'
      - description: routineFormIdId
        in: path
        name: routineFormIdId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - RoutineFormId
    put:
      parameters:
      - description: routineFormId
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RoutineFormId'
      - description: routineFormIdId
        in: path
        name: routineFormIdId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - RoutineFormId
...
  /users:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: user array
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: List
      tags:
      - User
    post:
      parameters:
      - description: User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - User
  /users/{userId}:
    delete:
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - User
    get:
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: user object
          schema:
            $ref: '#/definitions/models.User'
      summary: Get
      tags:
      - User
    patch:
      parameters:
      - description: user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - User
    put:
      parameters:
      - description: user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - User
...
swagger: "2.0"